+++ src/actions/index.js
@@ -11,47 +11,6 @@ export function resetData(data) {
   };
 }
 
-export const TOGGLE_NODE_CLICKED = 'TOGGLE_NODE_CLICKED';
-
-/**
- * Update the value of the currently-active clicked node
- * @param {string|null} nodeClicked The node's unique identifier
- */
-export function toggleNodeClicked(nodeClicked) {
-  return {
-    type: TOGGLE_NODE_CLICKED,
-    nodeClicked
-  };
-}
-
-export const TOGGLE_NODES_DISABLED = 'TOGGLE_NODES_DISABLED';
-
-/**
- * Toggle a selected group of nodes' visibility on/off
- * @param {Array} nodeIDs The nodes' unique identifiers
- * @param {Boolean} isDisabled Whether the node should be visible
- */
-export function toggleNodesDisabled(nodeIDs, isDisabled) {
-  return {
-    type: TOGGLE_NODES_DISABLED,
-    nodeIDs,
-    isDisabled
-  };
-}
-
-export const TOGGLE_NODE_HOVERED = 'TOGGLE_NODE_HOVERED';
-
-/**
- * Update the value of the currently-active hovered node
- * @param {string|null} nodeHovered The node's unique identifier
- */
-export function toggleNodeHovered(nodeHovered) {
-  return {
-    type: TOGGLE_NODE_HOVERED,
-    nodeHovered
-  };
-}
-
 export const TOGGLE_TEXT_LABELS = 'TOGGLE_TEXT_LABELS';
 
 /**
@@ -62,21 +21,6 @@ export function toggleTextLabels(textLabels) {
   return {
     type: TOGGLE_TEXT_LABELS,
     textLabels
-  };
-}
-
-export const TOGGLE_TAG_ACTIVE = 'TOGGLE_TAG_ACTIVE';
-
-/**
- * Toggle a tag's highlighting on/off
- * @param {string} tagID Tag id
- * @param {Boolean} active True if tag is active
- */
-export function toggleTagActive(tagID, active) {
-  return {
-    type: TOGGLE_TAG_ACTIVE,
-    tagID,
-    active
   };
 }
 
@@ -105,21 +49,6 @@ export function toggleTheme(theme) {
   return {
     type: TOGGLE_THEME,
     theme
-  };
-}
-
-export const TOGGLE_TYPE_DISABLED = 'TOGGLE_TYPE_DISABLED';
-
-/**
- * Toggle a node-type's visibility on/off
- * @param {string} typeID Type id
- * @param {Boolean} disabled True if type is disabled
- */
-export function toggleTypeDisabled(typeID, disabled) {
-  return {
-    type: TOGGLE_TYPE_DISABLED,
-    typeID,
-    disabled
   };
 }
