+++ src/components/tag-list/index.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import Checkbox from '@quantumblack/kedro-ui/lib/components/checkbox';
 import Dropdown from '@quantumblack/kedro-ui/lib/components/dropdown';
 import '@quantumblack/kedro-ui/lib/components/menu-option/menu-option.css';
-import { toggleTagActive, toggleTagFilter } from '../../actions';
+import { toggleTagActive, toggleTagFilter } from '../../actions/tags';
 import { getTagData, getTagCount } from '../../selectors/tags';
 import './tag-list.css';
